// src/types/tracks.ts
import { UserType } from "./user";

export type TrackType = {
  id: number;
  name: string;
  author: string;
  release_date: string;
  genre: string;
  duration_in_seconds: number;
  album: string;
  logo: string | null;
  track_file: string;
  stared_user: UserType[];
};

export type TrackListType = TrackType[];